Configure: allow to enable ktls if target does not start with Linux
The Debian build system uses a `debian' target which sets CFLAGS and
then we have for instance debian-amd64 which inherits from
linux-x86_64 and debian. So far so good.

Since the target name does not start with `linux', the build system does not
enable ktls. So in order to get enabled, I
added a
        `enable           => [ "ktls" ],'
to the generic linux config which sets it explicit). Having this set, we can
check for it instead matching the target name.

This commit is based on changes for afalgeng in commit
   9e381e8a01859 ("Configure: allow to enable afalgeng if target does not start with Linux")

Signed-off-by: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>

Gbp-Pq: Name Configure-allow-to-enable-ktls-if-target-does-not-start-w.patch